requirements.txt
pipecat-ai[google,silero]
fastapi
uvicorn
python-dotenv
python-multipart
loguru
twilio
